admin/pyproject.toml
[tool.poetry]
name = "bare-shelf-recipes-admin"
version = "0.1.0"
description = "The admin side for bareshelf.recipe"
authors = ["Rob Young <rob@robyoung.digital>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.1"
flask_admin = "^1.5.5"
flask-sqlalchemy = "^2.4.1"
psycopg2 = "^2.8.4"
Scrapy = "^2.0.1"
Flask-Migrate = "^2.5.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
